requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2; python_version >= "3.13"
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
faiss-cpu==1.12.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gradio==5.45.0
gradio_client==1.13.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.5
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.3.3
openai==1.107.2
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
psutil==7.1.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
ruff==0.13.0
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semantic-version==2.10.0
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
zstandard==0.25.0
